# PERF: use c-division in nancorr