- [ ] closes #xxxx (Replace xxxx with the Github issue number)
- [ ] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [ ] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [ ] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [ ] Added an entry in the latest `doc/source/whatsnew/vX.X.X.rst` file if fixing a bug or adding a new feature.

There's three divisions in this function that can safely be done at the C-level:

        meanx += 1 / nobs * dx
        meany += 1 / nobs * dy

because `nobs` is at least `1` when we get here (it starts at `0`, `nobs+=1` is called a few lines
above, and `nobs` can't decrease), and

        result[xi, yi] = result[yi, xi] = covxy / divisor

because `divisor` is checked to not be equal to `0` in the preceding line.

This makes a noticeable difference, example:
```
arr1 = np.random.randn(100, 100)
```
On main:
```
%%timeit
nancorr(arr1)

5.39 ms ± 21.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

On this branch:
```
%%timeit
nancorr(arr1)

3.86 ms ± 58.4 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

I didn't find much difference in nancorr_spearman, I think because that one calls
the Python function rank_1d within the loop, so that might be a bigger bottleneck there